+++ CHANGELOG.md
@@ -33,6 +33,7 @@ LCD and Math Controllers have been completely removed from Mycodo and replaced w
  - Add Input: TTN Integration: Data Storage (TTN v3, Payload jmespath Expression)
  - Add Output: Kasa HS300 Smart WiFi Power Strip
  - Add Output: I/O Expander: PCF8574 (8 Channels): Peristaltic Pump (Generic)
+ - Add Output: I/O Expander: MCP23017 (16 Channels): Peristaltic Pump (Generic)
  - Add size of file for last still and timelapse images
  - Add ability to duplicate Inputs
  - Add ability to invert the saved duty cycle for PWM Outputs